accel-config: Use SPDX identifiers in spec file
Update spec file to use SPDX identifiers for licenses, set correct
license value, and clean up the comments. The spec currently is
listing licenses of dependencies as well as the accel-config code,
where the License field should only show the license of the code being
shipped in the package. This was cleaned up in the upstream spec file
templates at some point, but never migrated to here. There is an
effort taking place now as well to move Fedora packages to using SPDX
identifiers for licenses, so make the switch here. A PR has also been
sent to the upstream project to make use of the SPDX identifiers in
the spec file templates there as well.

Signed-off-by: Jerry Snitselaar <jsnitsel@redhat.com>